# Add extra protections around wp_load_alloptions()