--- a/001-core.php
+++ b/001-core.php
@@ -9,6 +9,10 @@
 
 require_once __DIR__ . '/001-core/privacy.php';
 
+if ( file_exists( __DIR__ . '/001-core/options-api.php' ) ) {
+	require_once __DIR__ . '/001-core/options-api.php';
+}
+
 /**
  * Disable current theme validation
  *