--- a/001-core/options-api.php
+++ b/001-core/options-api.php
@@ -1,0 +1,65 @@
+<?php
+
+namespace Automattic\VIP\Core\OptionsAPI;
+
+/**
+ * Add additional protections around the alloptions functionality.
+ *
+ * Note that there is one (unavoidable) core call to get_option() before this filter is registered (in wp_plugin_directory_constants()),
+ * So by the time this starts filtering, there's already been one occurance of wp_load_alloptions().
+ *
+ * Here we re-implement most of what core does in wp_load_alloptions(), with some notable adjustments:
+ * - 1) Prevent spamming the DB if memcached is unable to add() the key to cache.
+ * - 2) (TBD) Kill the request if options cannot be retrieved from the database (or cache).
+ */
+add_filter( 'pre_wp_load_alloptions', __NAMESPACE__ . '\pre_wp_load_alloptions_protections', 2, 999 );
+function pre_wp_load_alloptions_protections( $pre_loaded_alloptions, $force_cache ) {
+	global $wpdb;
+	static $fallback_cache = [];
+
+	// Allow other filters the chance to early return (before priority 999).
+	// And abort from this special handling during installations.
+	if ( is_array( $pre_loaded_alloptions ) || wp_installing() ) {
+		return $pre_loaded_alloptions;
+	}
+
+	// 1) If successfully fetched from cache, return right away.
+	$alloptions_from_cache = wp_cache_get( 'alloptions', 'options', $force_cache );
+	if ( ! empty( $alloptions_from_cache ) && is_array( $alloptions_from_cache ) ) {
+		return apply_filters( 'alloptions', $alloptions_from_cache );
+	}
+
+	// 2) Return from the local fallback cache if available, helping prevent excess queries in cases where memcached is unable to save the results.
+	$blog_id = get_current_blog_id();
+	if ( ! $force_cache && isset( $fallback_cache[ $blog_id ] ) ) {
+		return apply_filters( 'alloptions', $fallback_cache[ $blog_id ] );
+	}
+
+	// 3) Otherwise query the DB for fresh results.
+	$suppress      = $wpdb->suppress_errors();
+	$alloptions_db = $wpdb->get_results( "SELECT option_name, option_value FROM $wpdb->options WHERE autoload = 'yes'" );
+	$wpdb->suppress_errors( $suppress );
+
+	$alloptions = [];
+	foreach ( (array) $alloptions_db as $o ) {
+		$alloptions[ $o->option_name ] = $o->option_value;
+	}
+
+	// TODO: Potentially kill the request rather than dangerously continue.
+	if ( empty( $alloptions ) ) {
+		trigger_error( 'VIP: Unable to query alloptions from database.', E_USER_WARNING );
+		return apply_filters( 'alloptions', apply_filters( 'pre_cache_alloptions', $alloptions ) );
+	}
+
+	$alloptions = apply_filters( 'pre_cache_alloptions', $alloptions );
+	$add_result = wp_cache_add( 'alloptions', $alloptions, 'options' );
+
+	if ( false === $add_result && false === wp_cache_get( 'alloptions', 'options', true ) ) {
+		// Start using the fallback cache if this request both failed to add() to cache, and there is
+		// nothing currently there - indicating there is likely something wrong with the ability to cache alloptions.
+		// Note that this is already the second time the request would have tried.
+		$fallback_cache[ $blog_id ] = $alloptions;
+	}
+
+	return apply_filters( 'alloptions', $alloptions );
+}