@WPprodigy 2023-07-21
Noting that this probably means we can start being less strict in https://github.com/Automattic/vip-go-mu-plugins/blob/c3be5f418a23223b89931d000528f5c312f55f09/alloptions-limit.php, and perhaps not force killing the request there. In fact, all of the mitigations there might be ripe for removal if/when we switch to individual cache keys for options.