@WPprodigy 2023-07-26
The `_VIP_DIE_ON_ALLOPTIONS_FAILURE` piece will be feature flagged and very selectively rolled out at the start. Can chat more about the other piece, but in short it's more or less what already happens with most other cache writes - in that it doesn't actually require success before the rest of the request starts serving it from runtime cache: https://github.com/Automattic/wp-cache-memcached/blob/develop/includes/wp-object-cache.php#L293-L301

The `add()` is just a bit different since it comes with the extra contract of having to be "first". In this alloptions case though, there are sets() that already happen after say an update_option(), so that contract isn't really in place. More specifically though, [because of this logic](https://github.com/Automattic/vip-go-mu-plugins/blob/develop/misc.php#L79-L96), any request could really be the "winner" for the add() race - including requests that are coming from a read replica. I have plans of solving for this too, but separately.

For what's in this PR, it's really just edge case protection to stop the DB and memcached from being spammed into an outage. And the divergent behavior introduced here only comes into play after two failed attempts to add() to memcached (so something is quite wrong). And the behavior isn't really different, because core would just keep querying the DB each time, attempting to add(), and then continuing with the results. So nothing changes with where or how it would have gotten the data, just saves lots of trouble.